Move RealTorManager to internal packaged file prefixed with -
